crawler/requirements.txt
kafka-python
requests
beautifulsoup4
selenium
webdriver-manager
kafka-handler
chromedriver-autoinstaller
